SERVER/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.1
pydantic-core==2.33.0
pymongo==4.11.3
pyparsing==3.2.3
pypdf2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
tqdm==4.67.1
typing-extensions==4.13.0
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.3.0
werkzeug==3.1.3
zipp==3.21.0
